// src/types/AppMode.ts
export type AppMode =
  | "home"
  | "specialist"
  | "conversation-menu"
  | "conversation"
  | "interview"
  | "quiz"
  | "games"
  | "listening"
  | "role-play"
  | "daily"
  | "assistant"
  | "free-talk"
  | "perguntas"
  | "content"
  | "flashcards-menu"
  | "flashcards"
  | "audio-flashcards"
  | "toefl"
  | "curso-completo"
  | "pronunciation"
  | "freePronunciation"
  | "tradutor";